Updates to RA control message

+++ model/satellite-phy-rx-carrier.cc
@@ -792,7 +792,7 @@ SatPhyRxCarrier::MeasureRandomAccessLoad ()
   NS_LOG_INFO ("SatPhyRxCarrier::MeasureRandomAccessLoad - Average normalized offered load: " << averageNormalizedOfferedLoad);
 
   /// TODO at the moment only one RA allocation channel is fully supported
-  uint32_t allocationChannelId = 0;
+  uint8_t allocationChannelId = 0;
 
   m_avgNormalizedOfferedLoadCallback (m_beamId, m_carrierId, allocationChannelId, averageNormalizedOfferedLoad);
 }
